home_page_templatetags moved to home_view
home_page_templatetags deleted, latest_published_tutorials and most_liked_tutorials handled in home_view instead
